+++ guillotina/tests/fixtures.py
@@ -118,14 +118,18 @@ def get_db_settings(pytest_node=None):
 
     settings["databases"]["db"]["dsn"] = {
         "scheme": "postgres",
-        "dbname": "guillotina",
+        "dbname": annotations.get("pg_db", "guillotina"),
         "user": "postgres",
         "host": annotations.get("pg_host", "localhost"),
         "port": annotations.get("pg_port", 5432),
         "password": "",
     }
 
-    options = dict(host=annotations.get("pg_host", "localhost"), port=annotations.get("pg_port", 5432))
+    options = dict(
+        host=annotations.get("pg_host", "localhost"),
+        port=annotations.get("pg_port", 5432),
+        dbname=annotations.get("pg_db", "guillotina"),
+    )
 
     if annotations["testdatabase"] == "cockroachdb":
         configure_db(settings["databases"]["db"], **options, user="root", storage="cockroach")